move matplotlib util functions to matrix_plotting and remove _matplotlib_backend
